add delete function with conection to DB
