+++ src/App.js
@@ -1,7 +1,7 @@
 import "./App.css";
 import Navbar from "./components/Navbar";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import Timesheet from "./pages/Timesheet";
+import Timesheets from "./pages/Timesheets";
 
 function App() {
   return (
@@ -11,7 +11,7 @@ function App() {
 
         <Routes>
           <Route element="Home page" path="/" exact />
-          <Route element={<Timesheet />} path="/timesheet" exact></Route>
+          <Route element={<Timesheets />} path="/timesheet" exact></Route>
           <Route element="Users Page" path="/users" exact />
           <Route element="Projects Page" path="/projects" exact />
           <Route element="Login page" path="/login" exact></Route>
